+++ Mlib/Images/CvSift/CvKeyPointsFilter.cpp
@@ -8,33 +8,33 @@ using namespace Mlib::ocv::KeyPointsFilter;
 void Mlib::ocv::KeyPointsFilter::removeDuplicated( std::vector<KeyPoint>& keypoints )
 {
     int i, j, n = (int)keypoints.size();
-    std::vector<int> kpidx(n);
-    std::vector<uint8_t> mask(n, (uint8_t)1);
+    std::vector<int> kpidx((size_t)n);
+    std::vector<uint8_t> mask((size_t)n, (uint8_t)1);
 
     for( i = 0; i < n; i++ )
-        kpidx[i] = i;
+        kpidx[(size_t)i] = i;
     std::sort(kpidx.begin(), kpidx.end());
     for( i = 1, j = 0; i < n; i++ )
     {
-        KeyPoint& kp1 = keypoints[kpidx[i]];
-        KeyPoint& kp2 = keypoints[kpidx[j]];
+        KeyPoint& kp1 = keypoints[(size_t)kpidx[(size_t)i]];
+        KeyPoint& kp2 = keypoints[(size_t)kpidx[(size_t)j]];
         if( kp1.pt(0) != kp2.pt(0) || kp1.pt(1) != kp2.pt(1) ||
             kp1.size != kp2.size || kp1.angle != kp2.angle )
             j = i;
         else
-            mask[kpidx[i]] = 0;
+            mask[(size_t)kpidx[(size_t)i]] = 0;
     }
 
     for( i = j = 0; i < n; i++ )
     {
-        if( mask[i] )
+        if( mask[(size_t)i] )
         {
             if( i != j )
-                keypoints[j] = keypoints[i];
+                keypoints[(size_t)j] = keypoints[(size_t)i];
             j++;
         }
     }
-    keypoints.resize(j);
+    keypoints.resize((size_t)j);
 }
 
 class MaskPredicate
@@ -95,13 +95,13 @@ void KeyPointsFilter::retainBest(std::vector<KeyPoint>& keypoints, int n_points)
         //first use nth element to partition the keypoints into the best and worst.
         std::nth_element(keypoints.begin(), keypoints.begin() + n_points - 1, keypoints.end(), KeypointResponseGreater());
         //this is the boundary response, and in the case of FAST may be ambiguous
-        float ambiguous_response = keypoints[n_points - 1].response;
+        float ambiguous_response = keypoints[size_t(n_points - 1)].response;
         //use std::partition to grab all of the keypoints with the boundary response.
         std::vector<KeyPoint>::const_iterator new_end =
         std::partition(keypoints.begin() + n_points, keypoints.end(),
                        KeypointResponseGreaterThanOrEqualToThreshold(ambiguous_response));
         //resize the keypoints, given this new end point. nth_element and partition reordered the points inplace
-        keypoints.resize(new_end - keypoints.begin());
+        keypoints.resize(size_t(new_end - keypoints.begin()));
 
         // std::sort(keypoints.begin(), keypoints.end(), KeypointResponseGreater());
     }
